+++ cmd/http-server/http-server.go
@@ -2,17 +2,10 @@ package main
 
 import (
 	"flag"
-	"fmt"
-	stdlog "log"
-	"net/http"
-	"os"
+	"log"
 
-	"github.com/go-kit/kit/log"
-
-	"github.com/julienschmidt/httprouter"
+	"github.com/livepeer/catalyst-api/api"
 	"github.com/livepeer/catalyst-api/config"
-	"github.com/livepeer/catalyst-api/handlers"
-	"github.com/livepeer/catalyst-api/middleware"
 	"github.com/livepeer/livepeer-data/pkg/mistconnector"
 )
 
@@ -27,36 +20,7 @@ func main() {
 		return
 	}
 
-	mc := &handlers.MistClient{
-		ApiUrl:          fmt.Sprintf("http://localhost:%d/api2", *mistPort),
-		TriggerCallback: fmt.Sprintf("http://localhost:%d/api/mist/trigger", *port),
+	if err := api.ListenAndServe(*port, *mistPort); err != nil {
+		log.Fatal(err)
 	}
-
-	listen := fmt.Sprintf("0.0.0.0:%d", *port)
-	router := NewCatalystAPIRouter(mc)
-
-	stdlog.Println("Starting Catalyst API version", config.Version, "listening on", listen)
-	err := http.ListenAndServe(listen, router)
-	stdlog.Fatal(err)
-
-}
-
-func NewCatalystAPIRouter(mc *handlers.MistClient) *httprouter.Router {
-	router := httprouter.New()
-
-	var logger log.Logger
-	logger = log.NewLogfmtLogger(log.NewSyncWriter(os.Stderr))
-	logger = log.With(logger, "ts", log.DefaultTimestampUTC)
-	withLogging := middleware.LogRequest(logger)
-
-	sc := make(map[string]handlers.StreamInfo)
-	catalystApiHandlers := &handlers.CatalystAPIHandlersCollection{MistClient: mc, StreamCache: sc}
-	mistCallbackHandlers := &handlers.MistCallbackHandlersCollection{MistClient: mc, StreamCache: sc}
-
-	router.GET("/ok", withLogging(catalystApiHandlers.Ok()))
-	router.POST("/api/vod", withLogging(middleware.IsAuthorized(catalystApiHandlers.UploadVOD())))
-	router.POST("/api/transcode/file", withLogging(middleware.IsAuthorized(catalystApiHandlers.TranscodeSegment())))
-	router.POST("/api/mist/trigger", withLogging(mistCallbackHandlers.Trigger()))
-
-	return router
 }
