+++ handlers/transcode.go
@@ -1,5 +1,15 @@
 package handlers
 
+import (
+	"encoding/json"
+	"io"
+	"net/http"
+
+	"github.com/julienschmidt/httprouter"
+	"github.com/livepeer/catalyst-api/errors"
+	"github.com/xeipuuv/gojsonschema"
+)
+
 type EncodedProfile struct {
 	Name         string `json:"name"`
 	Width        int    `json:"width"`
@@ -35,4 +45,35 @@ type TranscodeSegmentRequest struct {
 		} `json:"sceneClassification"`
 	} `json:"detection"`
 	VerificationFreq uint `json:"verificationFreq"`
+}
+
+func (d *CatalystAPIHandlersCollection) TranscodeSegment() httprouter.Handle {
+	schema := inputSchemasCompiled["TranscodeSegment"]
+
+	return func(w http.ResponseWriter, req *http.Request, _ httprouter.Params) {
+		var transcodeRequest TranscodeSegmentRequest
+		payload, err := io.ReadAll(req.Body)
+		if err != nil {
+			errors.WriteHTTPInternalServerError(w, "Cannot read body", err)
+			return
+		}
+		result, err := schema.Validate(gojsonschema.NewBytesLoader(payload))
+		if err != nil {
+			errors.WriteHTTPInternalServerError(w, "body schema validation failed", err)
+			return
+		}
+		if !result.Valid() {
+			errors.WriteHTTPBadBodySchema("TranscodeSegment", w, result.Errors())
+			return
+		}
+		if err := json.Unmarshal(payload, &transcodeRequest); err != nil {
+			errors.WriteHTTPBadRequest(w, "Invalid request payload", err)
+			return
+		}
+
+		if err := CallbackClient.SendTranscodeStatusError(transcodeRequest.CallbackUrl, "NYI - not yet implemented"); err != nil {
+			errors.WriteHTTPInternalServerError(w, "error send transcode error", err)
+		}
+		io.WriteString(w, "OK") // TODO later
+	}
 }
