+++ middleware/logging.go
@@ -5,8 +5,8 @@ import (
 	"runtime/debug"
 	"time"
 
-	log "github.com/go-kit/kit/log"
 	"github.com/julienschmidt/httprouter"
+	"github.com/livepeer/catalyst-api/config"
 	"github.com/livepeer/catalyst-api/errors"
 )
 
@@ -28,11 +28,9 @@ func (rw *responseWriter) WriteHeader(code int) {
 	rw.status = code
 	rw.ResponseWriter.WriteHeader(code)
 	rw.wroteHeader = true
-
-	return
 }
 
-func LogRequest(logger log.Logger) func(httprouter.Handle) httprouter.Handle {
+func LogRequest() func(httprouter.Handle) httprouter.Handle {
 	return func(next httprouter.Handle) httprouter.Handle {
 		fn := func(w http.ResponseWriter, r *http.Request, ps httprouter.Params) {
 			start := time.Now()
@@ -41,12 +39,12 @@ func LogRequest(logger log.Logger) func(httprouter.Handle) httprouter.Handle {
 			defer func() {
 				if err := recover(); err != nil {
 					errors.WriteHTTPInternalServerError(wrapped, "Internal Server Error", nil)
-					logger.Log("err", err, "trace", debug.Stack())
+					config.Logger.Log("err", err, "trace", debug.Stack())
 				}
 			}()
 
 			next(wrapped, r, ps)
-			logger.Log(
+			config.Logger.Log(
 				"remote", r.RemoteAddr,
 				"proto", r.Proto,
 				"method", r.Method,
